Move test registry data into test folder
