+++ ui/plugin-system/src/test/render.tsx
@@ -13,6 +13,8 @@
 
 import { render, RenderOptions } from '@testing-library/react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { PluginRegistry } from '../components/PluginRegistry';
+import { testRegistryProps } from './test-plugins';
 
 const testLogger = {
   log: console.log,
@@ -23,13 +25,19 @@ const testLogger = {
 };
 
 /**
- * Test helper to render a React component with some common app-level providers wrapped around it.
+ * Test helper to render a React component with some common app-level providers, as well as the PluginRegistry
+ * wrapped around it.
  */
-export function renderWithContext(ui: React.ReactElement, options?: Omit<RenderOptions, 'queries'>) {
+export function renderWithContext(ui: React.ReactNode, options?: Omit<RenderOptions, 'queries'>) {
   // Create a new QueryClient for each test to avoid caching issues
   const queryClient = new QueryClient({
     defaultOptions: { queries: { refetchOnWindowFocus: false, retry: false } },
     logger: testLogger,
   });
-  return render(<QueryClientProvider client={queryClient}>{ui}</QueryClientProvider>, options);
+  return render(
+    <QueryClientProvider client={queryClient}>
+      <PluginRegistry {...testRegistryProps}>{ui}</PluginRegistry>
+    </QueryClientProvider>,
+    options
+  );
 }
