+++ ui/plugin-system/src/test/test-plugins/bert/plugin.json
