+++ ui/plugin-system/src/test/test-plugins/ernie/plugin.json
